+++ integration/src/app/app.module.ts
@@ -4,7 +4,7 @@ import { FormsModule } from '@angular/forms';
 import { CommonModule } from '@angular/common';
 
 import { AppComponent } from './app.component';
-import { NgSelectModule, NG_SELECT_DEFAULT_CONFIG } from '@ng-select/ng-select';
+import { NgSelectModule } from '@ng-select/ng-select';
 
 @NgModule({
   declarations: [
@@ -13,16 +13,8 @@ import { NgSelectModule, NG_SELECT_DEFAULT_CONFIG } from '@ng-select/ng-select';
   imports: [
     CommonModule,
     FormsModule,
-    BrowserModule,
+    BrowserModule.withServerTransition({ appId: 'ng-select' }),
     NgSelectModule
-  ],
-  providers: [
-    {
-      provide: NG_SELECT_DEFAULT_CONFIG,
-      useValue: {
-        notFoundText: 'Custom not found'
-      }
-    }
   ],
   bootstrap: [AppComponent]
 })