+++ src/ng-select/ng-select.component.spec.ts
@@ -1,37 +1,15 @@
-import {
-    async,
-    ComponentFixture,
-    discardPeriodicTasks,
-    fakeAsync,
-    TestBed,
-    tick
-} from '@angular/core/testing';
+import { async, ComponentFixture, discardPeriodicTasks, fakeAsync, TestBed, tick } from '@angular/core/testing';
 import { By } from '@angular/platform-browser';
-import {
-    Component,
-    DebugElement,
-    ErrorHandler,
-    NgZone,
-    Type,
-    ViewChild
-} from '@angular/core';
+import { Component, DebugElement, ErrorHandler, NgZone, Type, ViewChild, ViewEncapsulation } from '@angular/core';
 import { ConsoleService } from './console.service';
 import { FormsModule } from '@angular/forms';
-import {
-    getNgSelectElement,
-    selectOption,
-    TestsErrorHandler,
-    tickAndDetectChanges,
-    triggerKeyDownEvent
-} from '../testing/helpers';
+import { getNgSelectElement, selectOption, TestsErrorHandler, tickAndDetectChanges, triggerKeyDownEvent } from '../testing/helpers';
 import { KeyCode, NgOption } from './ng-select.types';
 import { MockConsole, MockNgWindow, MockNgZone } from '../testing/mocks';
 import { NgSelectComponent } from './ng-select.component';
 import { NgSelectModule } from './ng-select.module';
 import { Subject } from 'rxjs';
 import { WindowService } from './window.service';
-
-
 
 describe('NgSelectComponent', function () {
 
@@ -357,6 +335,30 @@ describe('NgSelectComponent', function () {
             expect(fixture.componentInstance.selectedCity).toBeNull();
         }));
 
+        it('should map selected items with items in dropdown', fakeAsync(() => {
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities"
+                        bindLabel="name"
+                        [clearable]="true"
+                        [(ngModel)]="selectedCity">
+                </ng-select>`);
+
+            select = fixture.componentInstance.select;
+
+            fixture.componentInstance.selectedCity = fixture.componentInstance.cities[0];
+            tickAndDetectChanges(fixture);
+
+            fixture.componentInstance.cities = [...fixture.componentInstance.cities];
+            tickAndDetectChanges(fixture);
+
+            fixture.componentInstance.cities = [...fixture.componentInstance.cities];
+            tickAndDetectChanges(fixture);
+
+            expect(fixture.componentInstance.selectedCity).toEqual(fixture.componentInstance.cities[0]);
+            expect(select.itemsList.filteredItems[0].selected).toBeTruthy();
+        }));
+
         it('should clear previous single select value when setting new model', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
@@ -444,7 +446,7 @@ describe('NgSelectComponent', function () {
             expect(fixture.componentInstance.select.itemsList.markedItem.value).toEqual({ name: 'Vilnius', id: 1 });
         }));
 
-        it('bind to custom object properties', fakeAsync(() => {
+        it('should bind to custom object properties', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
                 `<ng-select [items]="cities"
@@ -453,12 +455,10 @@ describe('NgSelectComponent', function () {
                             [(ngModel)]="selectedCityId">
                 </ng-select>`);
 
-            // from component to model
             selectOption(fixture, KeyCode.ArrowDown, 0);
             tickAndDetectChanges(fixture);
             expect(fixture.componentInstance.selectedCityId).toEqual(1);
 
-            // from model to component
             fixture.componentInstance.selectedCityId = 2;
             tickAndDetectChanges(fixture);
             expect(fixture.componentInstance.select.selectedItems).toEqual([jasmine.objectContaining({
@@ -466,7 +466,7 @@ describe('NgSelectComponent', function () {
             })]);
         }));
 
-        it('bind to nested label property', fakeAsync(() => {
+        it('should bind to nested label property', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
                 `<ng-select [items]="countries"
@@ -489,7 +489,7 @@ describe('NgSelectComponent', function () {
             })]);
         }));
 
-        it('bind to nested value property', fakeAsync(() => {
+        it('should bind bind to nested value property', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
                 `<ng-select [items]="countries"
@@ -514,7 +514,7 @@ describe('NgSelectComponent', function () {
             expect(fixture.componentInstance.selectedCountry).toEqual('b');
         }));
 
-        it('bind to simple array', fakeAsync(() => {
+        it('should bind to simple array', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
                 `<ng-select [items]="citiesNames"
@@ -530,7 +530,7 @@ describe('NgSelectComponent', function () {
                 .toEqual([jasmine.objectContaining({ label: 'Kaunas', value: 'Kaunas' })]);
         }));
 
-        it('bind to object', fakeAsync(() => {
+        it('should bind to object', fakeAsync(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
                 `<ng-select [items]="cities"
@@ -964,6 +964,41 @@ describe('NgSelectComponent', function () {
             })
         }));
 
+        it('should select item with encapsulation = native', fakeAsync(() => {
+            const fixture = createTestingModule(
+                EncapsulatedTestCmp,
+                `<ng-select [items]="cities"
+                            bindLabel="name"
+                            [(ngModel)]="city"></ng-select>`);
+
+            expect((<NgOption[]>fixture.componentInstance.select.selectedItems).length).toBe(0);
+            expect(fixture.componentInstance.select.isOpen).toBeFalsy();
+
+            const cmp = fixture.componentInstance;
+
+            cmp.select.open();
+            tickAndDetectChanges(fixture);
+
+            expect((<NgOption[]>fixture.componentInstance.select.selectedItems).length).toBe(0);
+            expect(fixture.componentInstance.select.isOpen).toBeTruthy();
+
+            const outsideClick = spyOn(cmp.select.dropdownPanel.outsideClick, 'emit');
+            expect(outsideClick).not.toHaveBeenCalled();
+
+            const listItem = fixture.debugElement.query(By.css('.ng-option'));
+            let event = new MouseEvent('mousedown', { bubbles: true });
+            listItem.nativeElement.dispatchEvent(event);
+            event = new MouseEvent('click', { bubbles: true });
+            listItem.nativeElement.dispatchEvent(event);
+            tickAndDetectChanges(fixture);
+
+            fixture.whenStable().then(() => {
+                expect(outsideClick).not.toHaveBeenCalled();
+                expect(fixture.componentInstance.select.isOpen).toBeFalsy();
+                expect((<NgOption[]>fixture.componentInstance.select.selectedItems).length).toBe(1);
+            })
+        }));
+
         it('should not close when isOpen is true', async(() => {
             const fixture = createTestingModule(
                 NgSelectTestCmp,
@@ -1022,7 +1057,7 @@ describe('NgSelectComponent', function () {
 
             it('should not open dropdown when isOpen is false', () => {
                 const open = spyOn(select, 'open');
-                select.ngOnChanges(<any>{ isOpen: { currentValue: false }});
+                select.ngOnChanges(<any>{ isOpen: { currentValue: false } });
                 triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode.Space);
                 expect(select.isOpen).toBeFalsy();
                 expect(open).not.toHaveBeenCalled();
@@ -1236,6 +1271,18 @@ describe('NgSelectComponent', function () {
                 expect(select.selectedItems).toEqual(result);
                 expect(remove).toHaveBeenCalled();
             }));
+
+            it('should not remove selected value when clearOnBackspace false', fakeAsync(() => {
+                fixture.componentInstance.multiple = true;
+                select.clearOnBackspace = false;
+                fixture.componentInstance.cities = [...fixture.componentInstance.cities];
+                tickAndDetectChanges(fixture);
+                selectOption(fixture, KeyCode.ArrowDown, 1);
+                selectOption(fixture, KeyCode.ArrowDown, 1);
+                tickAndDetectChanges(fixture);
+                triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode.Backspace);
+                expect(select.selectedItems.length).toEqual(2);
+            }));
         });
 
         describe('key presses', () => {
@@ -1277,6 +1324,12 @@ describe('NgSelectComponent', function () {
                 expect(select.selectedItems[0].value).toEqual(fixture.componentInstance.cities[0]);
                 expect(select.isOpen).toBe(false);
             });
+
+            it('should not open dropdown if [openOnEnter]="false"', fakeAsync(() => {
+                select.openOnEnter = false;
+                triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode.Enter);
+                expect(select.isOpen).toBe(false);
+            }));
         });
     });
 
@@ -1804,6 +1857,38 @@ describe('NgSelectComponent', function () {
             expect(fixture.componentInstance.selectedCity).toBe(<any>'Copenhagen');
         }));
 
+        it('should add tag as string when there are no items', fakeAsync(() => {
+            let fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="[]"
+                    [addTag]="true"
+                    [(ngModel)]="selectedCity">
+                </ng-select>`);
+
+            tickAndDetectChanges(fixture);
+            fixture.componentInstance.select.filter('Copenhagen');
+            tickAndDetectChanges(fixture);
+            triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode.Enter);
+            expect(fixture.componentInstance.selectedCity).toBe(<any>'Copenhagen');
+            expect(fixture.componentInstance.select.itemsList.filteredItems.length).toBe(1);
+        }));
+
+        it('should not add item to list when select is closed', fakeAsync(() => {
+            let fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="[]"
+                    [isOpen]="false"
+                    [addTag]="true"
+                    [(ngModel)]="selectedCity">
+                </ng-select>`);
+
+            tickAndDetectChanges(fixture);
+            fixture.componentInstance.select.filter('Copenhagen');
+            tickAndDetectChanges(fixture);
+            triggerKeyDownEvent(getNgSelectElement(fixture), KeyCode.Enter);
+            expect(fixture.componentInstance.select.itemsList.filteredItems.length).toBe(0);
+        }));
+
         it('should add tag as string when tab pressed', fakeAsync(() => {
             let fixture = createTestingModule(
                 NgSelectTestCmp,
@@ -1910,9 +1995,19 @@ describe('NgSelectComponent', function () {
             });
 
             it('should be false when term exists among selected items', fakeAsync(() => {
-                fixture.componentInstance.selectedCities = [{name: 'Palanga', id: 9}];
+                fixture.componentInstance.selectedCities = [{ name: 'Palanga', id: 9 }];
                 select.filterValue = 'Palanga';
                 select.hideSelected = true;
+                select.isOpen = true;
+                tickAndDetectChanges(fixture);
+                expect(select.showAddTag).toBeFalsy();
+            }));
+
+            it('should be false when term exists among selected items and select is closed', fakeAsync(() => {
+                fixture.componentInstance.selectedCities = [{ name: 'Palanga', id: 9 }];
+                select.filterValue = 'Palanga';
+                select.hideSelected = false;
+                select.isOpen = false;
                 tickAndDetectChanges(fixture);
                 expect(select.showAddTag).toBeFalsy();
             }));
@@ -2359,7 +2454,7 @@ describe('NgSelectComponent', function () {
         });
     });
 
-    describe('Accessability', () => {
+    describe('Accessibility', () => {
         let fixture: ComponentFixture<NgSelectTestCmp>;
         let select: NgSelectComponent;
         let input: HTMLInputElement;
@@ -2610,6 +2705,23 @@ describe('NgSelectComponent', function () {
         }));
     });
 
+    describe('Auto-focus', () => {
+        it('should focus dropdown', fakeAsync(() => {
+            const fixture = createTestingModule(
+                NgSelectTestCmp,
+                `<ng-select [items]="cities"
+                            autofocus
+                            bindLabel="name"
+                            [multiple]="true"
+                            [(ngModel)]="selectedCities">
+                </ng-select>`);
+            const select = fixture.componentInstance.select;
+            const focus = spyOn(select, 'focus');
+            select.ngAfterViewInit();
+            expect(focus).toHaveBeenCalled();
+        }));
+    });
+
     describe('Mousedown', () => {
         let fixture: ComponentFixture<NgSelectTestCmp>;
         let select: NgSelectComponent;
@@ -2686,7 +2798,9 @@ describe('NgSelectComponent', function () {
                 tickAndDetectChanges(fixture);
                 triggerMousedown = () => {
                     const control = fixture.debugElement.query(By.css('.ng-select-container'));
-                    control.triggerEventHandler('mousedown', createEvent({ className: 'ng-clear' }));
+                    control.triggerEventHandler('mousedown', createEvent({
+                        classList: { contains: (term) => term === 'ng-clear-wrapper' }
+                    }));
                 };
             }));
 
@@ -2738,7 +2852,9 @@ describe('NgSelectComponent', function () {
                 tickAndDetectChanges(fixture);
                 triggerMousedown = () => {
                     const control = fixture.debugElement.query(By.css('.ng-select-container'));
-                    control.triggerEventHandler('mousedown', createEvent({ className: 'ng-value-icon' }));
+                    control.triggerEventHandler('mousedown', createEvent({
+                        classList: { contains: (term) => term === 'ng-value-icon' }
+                    }));
                 };
             }));
 
@@ -2769,7 +2885,9 @@ describe('NgSelectComponent', function () {
                 tickAndDetectChanges(fixture);
                 triggerMousedown = () => {
                     const control = fixture.debugElement.query(By.css('.ng-select-container'));
-                    control.triggerEventHandler('mousedown', createEvent({className: 'ng-arrow-wrapper' }));
+                    control.triggerEventHandler('mousedown', createEvent({
+                        classList: { contains: (term) => term === 'ng-arrow-wrapper' }
+                    }));
                 };
             }));
 
@@ -3028,6 +3146,9 @@ function createEvent(target = {}) {
         target: {
             className: '',
             tagName: '',
+            classList: {
+                contains: () => { }
+            },
             ...target
         }
     }
@@ -3039,6 +3160,7 @@ function createEvent(target = {}) {
 class NgSelectTestCmp {
     @ViewChild(NgSelectComponent) select: NgSelectComponent;
     multiple = false;
+    clearOnBackspace = false;
     disabled = false;
     dropdownPosition = 'bottom';
     visible = true;
@@ -3098,14 +3220,22 @@ class NgSelectTestCmp {
 }
 
 @Component({
-    template: ``
+    template: ``,
+    encapsulation: ViewEncapsulation.Native,
+})
+class EncapsulatedTestCmp extends NgSelectTestCmp {
+    @ViewChild(NgSelectComponent) select: NgSelectComponent;
+}
+
+@Component({
+    template: ``,
 })
 class NgSelectGroupingTestCmp {
     @ViewChild(NgSelectComponent) select: NgSelectComponent;
     selectedAccountName = 'Adam';
     selectedAccount = null;
     groupByFn = (item) => item.child.name;
-    groupValueFn = (key, _) => ({ group: key});
+    groupValueFn = (key, _) => ({ group: key });
     accounts = [
         { name: 'Adam', email: 'adam@email.com', age: 12, country: 'United States', child: { name: 'c1' } },
         { name: 'Samantha', email: 'samantha@email.com', age: 30, country: 'United States', child: { name: 'c1' } },